Add files via upload
Credit risk is one of the prominent
risks in banking and has a significant impact on the profitability rates of banks. Consequently,
developing credit scoring systems to assist banks and other financial institutions in making decisions
about lending to consumers is important. Financial institutions resort to various internal credit
evaluation models that reveal the relationships between factors thought to have an impact on credit
scores, in order to determine the path they will follow in granting credit to customers requesting
credit/debt. In the literature, statistics and machine learning techniques have been widely studied in
credit analysis. However, it is also believed that the developed software model will be beneficial to
organizations in increasing their profitability rates from loans.